REFACTOR: Make the matchmaking django project as single app
