--- services/matchmaking/main/templates/main/index.html
@@ -1,45 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <meta charset="utf-8"/>
-    <title>Chat Room</title>
-</head>
-<body>
-<!--    <textarea id="chat-log" cols="100" rows="20"></textarea><br>-->
-<!--    <input id="chat-message-input" type="text" size="100"><br>-->
-<!--    <input id="chat-message-submit" type="button" value="Send">-->
-    <script>
-
-        const chatSocket = new WebSocket(
-            'ws://'
-            + window.location.host
-            + '/ws/notification/'
-        );
-
-        chatSocket.onmessage = function(e) {
-            console.log(e.data);
-            // document.querySelector('#chat-log').value += (data.message + '\n');
-        };
-        //
-        // chatSocket.onclose = function(e) {
-        //     console.error('Chat socket closed unexpectedly');
-        // };
-        //
-        // document.querySelector('#chat-message-input').focus();
-        // document.querySelector('#chat-message-input').onkeyup = function(e) {
-        //     if (e.key === 'Enter') {  // enter, return
-        //         document.querySelector('#chat-message-submit').click();
-        //     }
-        // };
-        //
-        // document.querySelector('#chat-message-submit').onclick = function(e) {
-        //     const messageInputDom = document.querySelector('#chat-message-input');
-        //     const message = messageInputDom.value;
-        //     chatSocket.send(JSON.stringify({
-        //         'message': message
-        //     }));
-        //     messageInputDom.value = '';
-        // };
-    </script>
-</body>
-</html>
